Ensure plotted weeks start on a Monday.
All of this convoluted nonsense is to ensure that plotted weeks start on
Mondays. Otherwise, you get a problem generating plots toward the end of
the week - the first plotted week is partial and looks absurdly slow.

This little hack hopefully guarantees that each plotted week is a full week
*except* the most recent - and presumably people will be able to grok
why that one is slow.